Add plugin installation on opening sln * Add check if backup is available * Add missing path uproject file

+++ src/dotnet/RiderPlugin.UnrealLink/PluginInstaller/UnrealPluginInstaller.cs
@@ -66,11 +66,13 @@ namespace RiderPlugin.UnrealLink
                     }
                     catch (Exception _)
                     {
-                        backupDir.Move(pluginDir);
+                        if(backupDir.ExistsDirectory)
+                            backupDir.Move(pluginDir);
                         continue;
                     }
-
-                    backupDir.Delete();
+                    
+                    if(backupDir.ExistsDirectory)
+                        backupDir.Delete();
 
                     needToRegenerateProjectFiles = true;                    
                 }
